[DEBT-043-B] Split cli/scan.py into cli/scan/ package
- Split monolithic scan.py (1304 lines) into 7-file package (all ≤400 lines)
- Structure: __init__.py (113), _helpers.py (152), _opportunities_helpers.py (198),
  opportunities.py (341), new_markets.py (329), arbitrage.py (221), movers.py (215)
- Backward-compatible re-exports preserve existing test imports
- Tests updated, all 45 scan tests pass
- Quality gates pass: pre-commit, mypy, pytest (1014 tests)
- Acceptance criteria: Phase B complete
